Fix segementation fault caused by block notifies in quick succession
Since I was passing ownership of the StratumJob to the StratumClient unique_ptr, if another block.notify was issued soon after the first, it may be deleted by a second async thread before the first has finished operating on the StratumJob. To fix this, save the data we need later on before passing the StratumJob to the unique_ptr.